requirements.txt
# List of required Python packages.
twilio>=8.0.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
